feat: Allow users to left-click tab-completion popup items
Before this change, left-clicking an item in the tab-completion
popup would just highlight the item leaving the popup visible.
Left-clicking an item now selects that item to complete the current
command and hides the popup.

This commit also attempts to create a popup using the current
look and feel instead of just creating a BasicComboBox.

Test Plan:
```
1. Open a JRuby SwingReadline GUI.
2. Type `"foobar".`.
3. Hit the tab key.
4. Left-click an item in the combo box to select it.
```