feat: Make SwingConsoleFrame#run a non-blocking call
Since #run was blocking, clients would need to create their own
thread to not block the thread that calls #run. Since we already
create a thread, there is no reason for this method to block.

Also, the application does not immediately exit as long as the
AWT event queue is running. Applications naturally die once
the AWT event queue disposes of all windows.